카페 등록 추가(검증은 못 함)

+++ camoServer/src/main/java/jcw/camoServer/entity/Cafe.java
@@ -14,9 +14,9 @@ import javax.persistence.Id;
 @ToString
 public class Cafe {
     @Id
-    private int cafeId;
+    private Long cafeId;
 
-    private int memberId;
+    private Long memberId;
     private String cafeName;
     private String cafeAddress;
     private String cafeCallNumber;
